THREE.Legacy.js: Move deprecated propeties and methods to classes.
